[onert] Move EdgeTensor and PermuteFunction classes
This commit moves EdgeTensor and PermuteFunction classes to each class file to remove duplication.

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>
