+++ runtime/onert/core/src/exec/MultiModelExecutors.cc
@@ -58,77 +58,6 @@ namespace onert
 namespace exec
 {
 
-class MultiModelExecutors::EdgeTensor : public backend::IPortableTensor
-{
-public:
-  EdgeTensor(const ir::OperandInfo &info, ir::Layout layout)
-    : IPortableTensor(info), _layout{layout}, _buffer{nullptr}, _ref_count{0}
-  {
-  }
-  ~EdgeTensor() = default;
-
-  uint8_t *buffer() const override { return _buffer.get(); }
-  ir::Layout layout() const override { return _layout; }
-  void set_dynamic() override { _info.setDynamic(); }
-
-  // Almost same with basic::Tensor::applyShape() except dynamic memory allocation
-  bool applyShape(const ir::Shape &new_shape) override
-  {
-    bool previously_dynamic = is_dynamic();
-    if (!previously_dynamic || _buffer == nullptr)
-    {
-      // Always set shape - when buffer with same size was already allocated, shape could differ
-      setShape(new_shape);
-      set_dynamic();
-      const auto total_size = get_info().total_size();
-      _buffer = std::make_unique<uint8_t[]>(total_size);
-    }
-    else
-    {
-      auto previous_size = total_size();
-      auto new_size = new_shape.num_elements() * ir::sizeOfDataType(data_type());
-      if (previous_size != new_size)
-      {
-        setShape(new_shape);
-        set_dynamic();
-        const auto total_size = get_info().total_size();
-        _buffer = std::make_unique<uint8_t[]>(total_size);
-      }
-      else
-      { // when buffer with same size was already allocated, shape could differ
-        setShape(new_shape);
-      }
-    }
-    return true;
-  }
-
-  void setShape(const ir::Shape &new_shape) override { _info.shape(new_shape); }
-
-  void allocate_buffer()
-  {
-    const auto total_size = _info.total_size();
-    _buffer = std::make_unique<uint8_t[]>(total_size);
-    _ref_count = 1;
-  }
-
-  void increase_ref() { _ref_count++; }
-
-  void decrease_ref()
-  {
-    assert(_ref_count > 0);
-    _ref_count--;
-    if (_ref_count == 0)
-    {
-      _buffer.reset();
-    }
-  }
-
-private:
-  ir::Layout _layout;
-  std::unique_ptr<uint8_t[]> _buffer;
-  int32_t _ref_count;
-};
-
 void MultiModelExecutors::emplace(const ir::ModelIndex &model_index,
                                   const ir::SubgraphIndex &subg_index,
                                   std::unique_ptr<IExecutor> exec)
